Time: 4 ms (1.78%) | Memory: 8.2 MB (1.88%) - LeetSync

+++ 263-ugly-number/ugly-number.cpp
@@ -5,7 +5,7 @@ public:
         if(n<1){
             return false;
         }
-        for(long long i=2;i*i<=n;i++){
+        for(long i=2;i*i<=n;i++){
                while(n%i==0){
                    ans.push_back(i);
                    n=n/i;
